+++ fltk-sys/src/image.rs
@@ -71,6 +71,9 @@ extern "C" {
     pub fn Fl_Image_inactive(self_: *mut Fl_Image);
 }
 extern "C" {
+    pub fn Fl_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_Image;
+}
+extern "C" {
     pub fn Fl_Image_set_scaling_algorithm(algorithm: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -147,6 +150,9 @@ extern "C" {
     pub fn Fl_JPEG_Image_inactive(self_: *mut Fl_JPEG_Image);
 }
 extern "C" {
+    pub fn Fl_JPEG_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_JPEG_Image;
+}
+extern "C" {
     pub fn Fl_JPEG_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_JPEG_Image;
 }
 extern "C" {
@@ -221,6 +227,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_PNG_Image_inactive(self_: *mut Fl_PNG_Image);
+}
+extern "C" {
+    pub fn Fl_PNG_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_PNG_Image;
 }
 extern "C" {
     pub fn Fl_PNG_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_PNG_Image;
@@ -302,6 +311,9 @@ extern "C" {
     pub fn Fl_SVG_Image_inactive(self_: *mut Fl_SVG_Image);
 }
 extern "C" {
+    pub fn Fl_SVG_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_SVG_Image;
+}
+extern "C" {
     pub fn Fl_SVG_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_SVG_Image;
 }
 extern "C" {
@@ -379,6 +391,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_BMP_Image_inactive(self_: *mut Fl_BMP_Image);
+}
+extern "C" {
+    pub fn Fl_BMP_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_BMP_Image;
 }
 extern "C" {
     pub fn Fl_BMP_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_BMP_Image;
@@ -460,6 +475,9 @@ extern "C" {
     pub fn Fl_GIF_Image_inactive(self_: *mut Fl_GIF_Image);
 }
 extern "C" {
+    pub fn Fl_GIF_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_GIF_Image;
+}
+extern "C" {
     pub fn Fl_GIF_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_GIF_Image;
 }
 extern "C" {
@@ -539,6 +557,9 @@ extern "C" {
     pub fn Fl_Pixmap_inactive(self_: *mut Fl_Pixmap);
 }
 extern "C" {
+    pub fn Fl_Pixmap_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_Pixmap;
+}
+extern "C" {
     pub fn Fl_Pixmap_new(D: *const *const ::std::os::raw::c_char) -> *mut Fl_Pixmap;
 }
 #[repr(C)]
@@ -610,6 +631,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_XPM_Image_inactive(self_: *mut Fl_XPM_Image);
+}
+extern "C" {
+    pub fn Fl_XPM_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_XPM_Image;
 }
 extern "C" {
     pub fn Fl_XPM_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_XPM_Image;
@@ -685,6 +709,9 @@ extern "C" {
     pub fn Fl_XBM_Image_inactive(self_: *mut Fl_XBM_Image);
 }
 extern "C" {
+    pub fn Fl_XBM_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_XBM_Image;
+}
+extern "C" {
     pub fn Fl_XBM_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_XBM_Image;
 }
 #[repr(C)]
@@ -758,6 +785,9 @@ extern "C" {
     pub fn Fl_PNM_Image_inactive(self_: *mut Fl_PNM_Image);
 }
 extern "C" {
+    pub fn Fl_PNM_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_PNM_Image;
+}
+extern "C" {
     pub fn Fl_PNM_Image_new(filename: *const ::std::os::raw::c_char) -> *mut Fl_PNM_Image;
 }
 #[repr(C)]
@@ -829,6 +859,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Tiled_Image_inactive(self_: *mut Fl_Tiled_Image);
+}
+extern "C" {
+    pub fn Fl_Tiled_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_Tiled_Image;
 }
 extern "C" {
     pub fn Fl_Tiled_Image_new(
@@ -906,6 +939,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_RGB_Image_inactive(self_: *mut Fl_RGB_Image);
+}
+extern "C" {
+    pub fn Fl_RGB_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_RGB_Image;
 }
 extern "C" {
     pub fn Fl_RGB_Image_new(
@@ -1001,6 +1037,9 @@ extern "C" {
     pub fn Fl_Shared_Image_inactive(self_: *mut Fl_Shared_Image);
 }
 extern "C" {
+    pub fn Fl_Shared_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_Shared_Image;
+}
+extern "C" {
     pub fn Fl_Shared_Image_get(
         name: *const ::std::os::raw::c_char,
         W: ::std::os::raw::c_int,
@@ -1012,6 +1051,98 @@ extern "C" {
         rgb: *mut Fl_RGB_Image,
         own_it: ::std::os::raw::c_int,
     ) -> *mut Fl_Shared_Image;
+}
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub struct Fl_ICO_Image {
+    _unused: [u8; 0],
+}
+extern "C" {
+    pub fn Fl_ICO_Image_draw(
+        arg1: *mut Fl_ICO_Image,
+        X: ::std::os::raw::c_int,
+        Y: ::std::os::raw::c_int,
+        W: ::std::os::raw::c_int,
+        H: ::std::os::raw::c_int,
+    );
+}
+extern "C" {
+    pub fn Fl_ICO_Image_draw_ext(
+        arg1: *mut Fl_ICO_Image,
+        X: ::std::os::raw::c_int,
+        Y: ::std::os::raw::c_int,
+        W: ::std::os::raw::c_int,
+        H: ::std::os::raw::c_int,
+        cx: ::std::os::raw::c_int,
+        cy: ::std::os::raw::c_int,
+    );
+}
+extern "C" {
+    pub fn Fl_ICO_Image_width(arg1: *mut Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_height(arg1: *mut Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_delete(arg1: *mut Fl_ICO_Image);
+}
+extern "C" {
+    pub fn Fl_ICO_Image_count(self_: *mut Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_data(self_: *mut Fl_ICO_Image) -> *const *const ::std::os::raw::c_char;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_copy(self_: *mut Fl_ICO_Image) -> *mut Fl_ICO_Image;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_scale(
+        self_: *mut Fl_ICO_Image,
+        width: ::std::os::raw::c_int,
+        height: ::std::os::raw::c_int,
+        proportional: ::std::os::raw::c_int,
+        can_expand: ::std::os::raw::c_int,
+    );
+}
+extern "C" {
+    pub fn Fl_ICO_Image_fail(self_: *mut Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_data_w(self_: *const Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_data_h(self_: *const Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_d(self_: *const Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_ld(self_: *const Fl_ICO_Image) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_inactive(self_: *mut Fl_ICO_Image);
+}
+extern "C" {
+    pub fn Fl_ICO_Image_from_dyn_ptr(other: *mut Fl_Image) -> *mut Fl_ICO_Image;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_new(
+        filename: *const ::std::os::raw::c_char,
+        id: ::std::os::raw::c_int,
+    ) -> *mut Fl_ICO_Image;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_from_data(
+        bits: *const ::std::os::raw::c_uchar,
+        len: ::std::os::raw::c_ulong,
+        id: ::std::os::raw::c_int,
+    ) -> *mut Fl_ICO_Image;
+}
+extern "C" {
+    pub fn Fl_ICO_Image_icondirentry(
+        self_: *const Fl_ICO_Image,
+        size: *mut ::std::os::raw::c_ulong,
+    ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_register_images();