+++ fltk/src/macros/image.rs
@@ -86,7 +86,7 @@ macro_rules! impl_image_ext {
                     unsafe { [<$flname _height>](self.inner) }
                 }
 
-                unsafe fn as_image_ptr(&self) -> *mut fltk_sys::image::Fl_Image {
+                fn as_image_ptr(&self) -> *mut fltk_sys::image::Fl_Image {
                     assert!(!self.was_deleted());
                     self.inner as *mut fltk_sys::image::Fl_Image
                 }
@@ -211,6 +211,24 @@ macro_rules! impl_image_ext {
 
                 unsafe fn into_image<I: ImageExt>(self) -> I {
                     I::from_image_ptr(self.inner as *mut _)
+                }
+
+                fn from_dyn_image_ptr(p: *mut fltk_sys::image::Fl_Image) -> Option<Self> {
+                    unsafe {
+                        let ptr = [<$flname _from_dyn_ptr>](p);
+                        if ptr.is_null() {
+                            None
+                        } else {
+                            Some($name {
+                                inner: ptr as *mut $flname,
+                                refcount: std::sync::atomic::AtomicUsize::new(1),
+                            })
+                        }
+                    }
+                }
+
+                fn from_dyn_image<I: ImageExt>(i: &I) -> Option<Self> {
+                    Self::from_dyn_image_ptr(i.as_image_ptr())
                 }
             }
         }