+++ fltk/src/macros/widget.rs
@@ -316,7 +316,7 @@ macro_rules! impl_widget_ext {
                     (x, y)
                 }
 
-                unsafe fn as_widget_ptr(&self) -> *mut fltk_sys::widget::Fl_Widget {
+                fn as_widget_ptr(&self) -> *mut fltk_sys::widget::Fl_Widget {
                     self.inner as *mut fltk_sys::widget::Fl_Widget
                 }
 
@@ -728,10 +728,23 @@ macro_rules! impl_widget_ext {
                         if image_ptr.is_null() {
                             None
                         } else {
-                            let img =
-                            $crate::image::Image::from_image_ptr(image_ptr as *mut fltk_sys::image::Fl_Image);
-                            Some(Box::new(img))
+                            let mut img =
+                                $crate::image::Image::from_image_ptr(image_ptr as *mut fltk_sys::image::Fl_Image);
+                            img.increment_arc();
+                            Some(Box::new(img.copy()))
                         }
+                    }
+                }
+
+                unsafe fn image_mut(&self) -> Option<&mut $crate::image::Image> {
+                    assert!(!self.was_deleted());
+                    let image_ptr = [<$flname _image>](self.inner);
+                    if image_ptr.is_null() {
+                        None
+                    } else {
+                        let img =
+                            $crate::image::Image::from_image_ptr(image_ptr as *mut fltk_sys::image::Fl_Image);
+                        Some(Box::leak(Box::new(img)))
                     }
                 }
 
@@ -783,10 +796,23 @@ macro_rules! impl_widget_ext {
                         if image_ptr.is_null() {
                             None
                         } else {
-                            let img =
-                            $crate::image::Image::from_image_ptr(image_ptr as *mut fltk_sys::image::Fl_Image);
-                            Some(Box::new(img))
+                            let mut img =
+                                $crate::image::Image::from_image_ptr(image_ptr as *mut fltk_sys::image::Fl_Image);
+                            img.increment_arc();
+                            Some(Box::new(img.copy()))
                         }
+                    }
+                }
+
+                unsafe fn deimage_mut(&self) -> Option<&mut $crate::image::Image> {
+                    assert!(!self.was_deleted());
+                    let image_ptr = [<$flname _deimage>](self.inner);
+                    if image_ptr.is_null() {
+                        None
+                    } else {
+                        let img =
+                            $crate::image::Image::from_image_ptr(image_ptr as *mut fltk_sys::image::Fl_Image);
+                        Some(Box::leak(Box::new(img)))
                     }
                 }
 
@@ -835,7 +861,7 @@ macro_rules! impl_widget_ext {
                 }
 
                 fn is_same<W: WidgetExt>(&self, other: &W) -> bool {
-                    unsafe { self.as_widget_ptr() == other.as_widget_ptr() }
+                    self.as_widget_ptr() == other.as_widget_ptr()
                 }
 
                 fn active(&self) -> bool {
@@ -1088,6 +1114,27 @@ macro_rules! impl_widget_base {
                 unsafe fn assume_derived(&mut self) {
                     self.is_derived = true
                 }
+
+                fn from_dyn_widget<W: WidgetExt>(w: &W) -> Option<Self> {
+                    Self::from_dyn_widget_ptr(w.as_widget_ptr() as _)
+                }
+
+                fn from_dyn_widget_ptr(w: *mut fltk_sys::widget::Fl_Widget) -> Option<Self> {
+                    let ptr = unsafe { [<$flname _from_dyn_ptr>](w as _) };
+                    if ptr.is_null() {
+                        None
+                    } else {
+                        let tracker = unsafe {
+                            fltk_sys::fl::Fl_Widget_Tracker_new(ptr as *mut fltk_sys::fl::Fl_Widget)
+                        };
+                        assert!(!tracker.is_null());
+                        Some(Self {
+                            inner: ptr as *mut $flname,
+                            tracker,
+                            is_derived: false,
+                        })
+                    }
+                }
             }
         }
     };
@@ -1251,7 +1298,7 @@ macro_rules! impl_widget_ext_via {
                 self.$member.measure_label()
             }
 
-            unsafe fn as_widget_ptr(&self) -> $crate::app::WidgetPtr {
+            fn as_widget_ptr(&self) -> $crate::app::WidgetPtr {
                 self.$member.as_widget_ptr()
             }
 
@@ -1279,6 +1326,10 @@ macro_rules! impl_widget_ext_via {
                 self.$member.image()
             }
 
+            unsafe fn image_mut(&self) -> Option<&mut $crate::image::Image> {
+                self.$member.image_mut()
+            }
+
             fn set_deimage<I: ImageExt>(&mut self, image: Option<I>) {
                 self.$member.set_deimage(image)
             }
@@ -1289,6 +1340,10 @@ macro_rules! impl_widget_ext_via {
 
             fn deimage(&self) -> Option<Box<dyn ImageExt>> {
                 self.$member.deimage()
+            }
+
+            unsafe fn deimage_mut(&self) -> Option<&mut $crate::image::Image> {
+                self.$member.deimage_mut()
             }
 
             fn set_callback<F: FnMut(&mut Self) + 'static>(&mut self, mut cb: F) {
@@ -1623,6 +1678,22 @@ macro_rules! impl_widget_base_via {
 
             unsafe fn assume_derived(&mut self) {
                 self.assume_derived()
+            }
+
+            fn from_dyn_widget<W: WidgetExt>(w: &W) -> Self {
+                let $member = <$base>::from_dyn_widget(ptr);
+                Self {
+                    $member,
+                    ..Default::default()
+                }
+            }
+
+            fn from_dyn_widget_ptr(w: *mut fltk_sys::widget::Fl_Widget) -> Self {
+                let $member = <$base>::from_dyn_widget(ptr);
+                Self {
+                    $member,
+                    ..Default::default()
+                }
             }
         }
     };