+++ fltk/src/utils/mod.rs
@@ -129,3 +129,13 @@ pub fn content_view<W: crate::prelude::WindowExt>(w: &W) -> *const raw::c_void {
     }
     unsafe { cfltk_getContentView(w.raw_handle() as _) as _ }
 }
+
+/// Check whether a widget is of a certain type
+pub fn is<W: crate::prelude::WidgetBase>(w: &impl crate::prelude::WidgetBase) -> bool {
+    W::from_dyn_widget(w).is_some()
+}
+
+/// Check whether a widget is of a certain type
+pub fn is_ptr_of<W: crate::prelude::WidgetBase>(w: *mut fltk_sys::widget::Fl_Widget) -> bool {
+    W::from_dyn_widget_ptr(w).is_some()
+}